+++ ecoscope/base/__init__.py
@@ -5,25 +5,20 @@ from ecoscope.base._dataclasses import (
     RelocsSpeedFilter,
     TrajSegFilter,
 )
-from ecoscope.base.new_base import NewRelocations
-from ecoscope.base.base import EcoDataFrame, Relocations, Trajectory
+from ecoscope.base.base import Relocations, Trajectory
 from ecoscope.base.utils import (
-    cachedproperty,
     create_meshgrid,
     groupby_intervals,
 )
 
 __all__ = [
-    "EcoDataFrame",
     "Relocations",
-    "NewRelocations",
     "RelocsCoordinateFilter",
     "RelocsDateRangeFilter",
     "RelocsDistFilter",
     "RelocsSpeedFilter",
     "TrajSegFilter",
     "Trajectory",
-    "cachedproperty",
     "create_meshgrid",
     "groupby_intervals",
 ]
